pyproject.toml
[project]
name = "papercheck"
version = "0.1.0"
description = "Multi-layer research paper verification pipeline"
requires-python = ">=3.11"

dependencies = [
    "click>=8.1",
    "pydantic>=2.5",
    "pymupdf>=1.23",
    "pylatexenc>=2.10",
    "lxml>=5.0",
    "anthropic>=0.40",
    "httpx>=0.27",
    "tenacity>=8.2",
    "diskcache>=5.6",
    "jinja2>=3.1",
    "rich>=13.0",
    "regex>=2023.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.4",
]
corpus = [
    "chromadb>=0.4",
    "sentence-transformers>=2.3",
]

[project.scripts]
papercheck = "papercheck.cli:main"

[tool.setuptools.packages.find]
include = ["papercheck*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["papercheck/tests"]
markers = [
    "live: marks tests that require network access and API keys (deselect with '-m not live')",
]

[tool.ruff]
target-version = "py311"
line-length = 100
